backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Pydantic
pydantic==2.4.2
pydantic-settings==2.0.3

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Pipecat for AI services (existing bot logic)
pipecat-ai[cartesia,openai,silero,deepgram]
pipecat

# Gemini AI (for existing service)
google-generativeai==0.3.2
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.16.3

# Twilio for call management
twilio==8.10.0

# HTTP requests
requests==2.31.0
httpx==0.25.2

# Logging
loguru>=0.7.0

# CORS and middleware
fastapi-cors==0.0.6

# Date and time
python-dateutil==2.8.2

# CSV processing
pandas==2.1.4

# WebSocket support
websockets==12.0

# Email validation
email-validator==2.1.0

# Environment variables
python-dotenv==1.0.0 